ADD: merging priyanshu's code
