+++ src/components/Contests.js
@@ -1,0 +1,49 @@
+import React, { useState, useEffect } from "react";
+import "./style.css";
+
+const Contests = () => {
+  const [contests, setContests] = useState([]);
+  // console.log(contests);
+
+  useEffect(() => {
+    const getContests = async (siteName) => {
+      await fetch("https://kontests.net/api/v1/" + siteName)
+        .then((response) => response.json())
+        .then((data) => {
+          // console.log(data);
+          const cntst = data.map((contest) => ({
+            name: contest.name,
+            link: contest.url,
+          }));
+          setContests(cntst);
+        });
+    };
+    getContests("codeforces");
+  }, []);
+
+  return (
+    <div className="contests">
+      <div className="d-flex justify-content-center">
+        <div>
+          {contests.map(({ name, link }) => {
+            return (
+              <tr key={link}>
+                <td>
+                  <ul className="text-center">
+                    <strong>
+                      <a href={link} target="_blank" rel="noreferrer">
+                        {name}
+                      </a>
+                    </strong>
+                  </ul>
+                </td>
+              </tr>
+            );
+          })}
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Contests;
