--- src/components/Home.css
@@ -1,3 +1,0 @@
-.SSDC{
-    margin-top: 165px;
-}
