+++ src/components/Home.js
@@ -3,22 +3,18 @@ import Contests from './Contests'
 import Navbar from './Navbar'
 import TextBox from './TextBox'
 import Footer from './Footer'
-import './Home.css'
+import './style.css'
 
 const Home = () => {
   return (
-    <div>
-      <Navbar />
-
-      <div className="SSDC">
+      <div className="SSDC bg">
+        <Navbar />
         <h1 className='text-center display-1'>SSDC</h1>
         <h3 className='text-center display-5'>Hive of Developers</h3>
+        <TextBox />
+        <Contests className="my-3"/>
+        <Footer />
       </div>
-
-      <TextBox />
-      <Contests />
-      <Footer />
-    </div>
   )
 }
