--- src/components/Navbar.css
@@ -1,49 +1,0 @@
-.nav{
-    position: fixed;
-    top: 0px;
-    width: 100%;
-    height: 90px;
-    z-index: 1;
-
-    display: flex;
-    align-items: center;
-    /* Animation */
-    transition-timing-function: ease-in;
-    transition: all 0.5s;
-}
-
-.nav__black{
-    background-color: black;
-    color: white;
-}
-
-.nav__logo{
-    position: fixed;
-    left: 0px;
-    width: 55px;
-    cursor: pointer;
-    margin-left: 20px;
-    border-radius: 50%;
-    object-fit: contain;
-}
-
-.nav__avatar{
-    cursor: pointer;
-    position: fixed;
-    right: 20px;
-    width: 30px;
-}
-
-.nav__elements{
-    width: 100%;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    margin-right: 60px;
-}
-
-.nav__element{
-    padding: 10px;
-    cursor: pointer;
-    letter-spacing: 1.42px;
-}
