+++ src/components/TextBox.js
@@ -2,30 +2,13 @@ import React from 'react'
 
 function TextBox() {
   return (
-    <div className='container my-3 box'>
-        <p className="text-in-box">Lorem ipsum dolor sit, amet consectetur adipisicing elit. Quis architecto velit nobis, corrupti dolore magnam exercitationem eligendi minus molestias beatae maxime ullam sapiente quisquam quas consequuntur eos quasi atque autem.
-        Unde inventore, quia eligendi velit quod voluptatum perspiciatis sit itaque minus cum, sapiente quos consectetur optio veritatis! Eos quod ullam assumenda? Cupiditate ad saepe ratione nemo aspernatur quia explicabo blanditiis.
-        Autem architecto sunt sint pedit officiis est corrupti, ex alias sequi quam ipsum deserunt facilis eveniet eius itaque sit excepturi. Atque odit commodi deserunt ullam, voluptates ipsam?
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Tenetur, error voluptas. Autem magni eligendi expedita quae? Eius non modi porro quibusdam harum. Ratione excepturi tempore nam quia, nihil mollitia minus.
-        Quidem vel omnis earum repudiandae rem, perspiciatis quis ad explicabo veniam sunt esse, recusandae vitae veritatis aut sed incidunt nisi quam molestiae doloremque consequatur non perferendis. Saepe fuga molestias nesciunt.
-        Quis, tempore rem amet mollitia vero nisi esse error nostrum non quia ea labore illum itaque delectus, est minus praesentium magnam sit ipsum molestias natus temporibus, voluptatem porro? Perferendis, id?
-        Error unde obcaecati ducimus, ut perspiciatis iste deleniti natus perferendis veniam expedita explicabo voluptatibus doloribus tempore, aperiam laboriosam, voluptate reiciendis hic itaque deserunt similique nihil odio consequatur vel nostrum! Porro.
-        Hic, eius sint? Mollitia ullam similique iure facilis odit dolore quibusdam nihil aperiam, culpa non, dicta laudantium sed porro aliquam minima doloribus! Minima molestias suscipit, sapiente cumque accusantium doloribus veniam?
-        Facilis porro ducimus similique ipsum quam molestiae officia corrupti nobis. Cumque ducimus harum distinctio iure incidunt? Tenetur, laborum at! Soluta, aliquid sapiente! Ipsam aut explicabo asperiores illo? Voluptate, voluptates nihil!
-        Aliquid labore dolorem dolore iusto incidunt consequuntur odit, culpa similique accusantium, ad reiciendis exercitationem voluptate? Ad repudiandae eius, enim laudantium, perferendis debitis unde officia numquam tempora eveniet consequuntur quam laborum.
-        Doloremque provident a quam blanditiis animi distinctio libero, veritatis odio quas adipisci at iusto beatae perspiciatis eius amet sed saepe in voluptatibus quo nisi laboriosam? Velit, repellat explicabo? Dolores, iure.
-        Ratione aspernatur totam quos eos, dolore cum illum voluptatum quae ad est, facere quia ex dicta, inventore voluptate! Commodi minus voluptatum iste quis earum pariatur dolorum qui, cum numquam perspiciatis.
-        Obcaecati dolore rerum, nobis debitis eius exercitationem eos suscipit dolor consequatur nemo beatae amet, perferendis quas impedit pariatur soluta vel quisquam recusandae libero cupiditate. Est ipsa totam doloribus vel magnam.
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Perspiciatis dolore adipisci, placeat possimus a aperiam architecto corporis, voluptatibus maxime, ratione ipsam hic iure culpa libero natus debitis dolores maiores veniam!
-        Quis aut reiciendis maiores commodi dolore voluptas debitis inventore, rem ab eum vitae voluptatibus recusandae, eius a perspiciatis nisi unde facere velit ea amet incidunt eaque nesciunt aperiam! At, fugiat!
-        Maxime perferendis facere expedita voluptatibus quas ipsa. Sit deserunt accusamus odit id eveniet quis ipsa, nemo itaque possimus et, cum, beatae repudiandae exercitationem rem veniam consectetur eaque! Quo, rerum earum.
-        Quis deleniti amet fugiat sed rem ad accusamus ullam fuga aliquid libero. Nulla maiores aliquid quasi voluptates iusto molestias minima cum sint hic, reprehenderit vitae commodi autem soluta deserunt facere.
-        Nobis hic architecto vero illum voluptatibus at quaerat tempora suscipit unde, explicabo commodi. Tenetur sit porro tempora inventore quam placeat culpa eaque, repudiandae facilis asperiores mollitia aspernatur suscipit atque maxime.
-        In maiores similique perferendis error sunt qui sint iste provident distinctio eaque deserunt a optio rerum possimus at, atque minus alias veniam commodi quibusdam molestiae animi corporis esse! Nostrum, a.
-        Eius reiciendis voluptatum nesciunt, voluptatem dicta modi laborum tempora officia illo consequuntur dolorem sed vel. Earum quis culpa voluptatem voluptatum, eligendi suscipit ratione magnam reprehenderit laboriosam, nam similique, maxime omnis.
-        Reiciendis laborum, delectus rem commodi expedita eaque possimus vero nihil, ullam ea veniam dicta a neque rerum distinctio aspernatur, est odio. Delectus animi quasi doloribus officia quibusdam exercitationem ipsa a!
-        Sequi laudantium ullam perferendis fuga dolorum dolorem ipsum impedit nam repellat tenetur, sed nemo obcaecati tempora error rem eveniet praesentium officiis labore voluptatem voluptas dolore quae magni! Dignissimos, quo quidem?
-        Vitae natus placeat, labore aliquid amet quis nesciunt ipsam, unde quo velit doloremque recusandae, nam voluptate debitis tempora? Hic sed commodi, eos excepturi expedita voluptates ad suscipit placeat aperiam fugiat!</p>
+    <div className='container my-3 box text-center'>
+        <p className="text-in-box">SLIET Software Development Club (SSDC), is a Departmental Club of Computer Science & Engineering, SLIET Longowal. This club was founded by CSE - 2017 undergrads.</p>
+        <br /><br />
+        <p>SLIET Software Development Club (SSDC), is a Departmental Club of Computer Science & Engineering, SLIET Longowal. This club was founded by CSE - 2017 undergrads. The team currently comprises of 40+ developers. The team aims to brush up the coding skills in young enthusiasts, with major goals to crack the latest Coding Contests, Challenges & Interviews without any reluctance landing them in some decent technical job prospects. In due course of grooming it's pupils, the club focuses on project-oriented approach in diverse (Web Development, App Development, Machine Learning, Cybersecurity, etc) to ensure holistic development of it's members. Furthermore, the club organises workshops and monthly coding compitions for college students as well which in turn helps us discover young wizards in IT fields.</p>
+        <p className="text-in-box">SLIET Software Development Club (SSDC), is a Departmental Club of Computer Science & Engineering, SLIET Longowal. This club was founded by CSE - 2017 undergrads.</p>
+        <br /><br />
+        <p>SLIET Software Development Club (SSDC), is a Departmental Club of Computer Science & Engineering, SLIET Longowal. This club was founded by CSE - 2017 undergrads. The team currently comprises of 40+ developers. The team aims to brush up the coding skills in young enthusiasts, with major goals to crack the latest Coding Contests, Challenges & Interviews without any reluctance landing them in some decent technical job prospects. In due course of grooming it's pupils, the club focuses on project-oriented approach in diverse (Web Development, App Development, Machine Learning, Cybersecurity, etc) to ensure holistic development of it's members. Furthermore, the club organises workshops and monthly coding compitions for college students as well which in turn helps us discover young wizards in IT fields.</p>
     </div>
   )
 }
